+++ example/lib/main.dart
@@ -12,8 +12,10 @@ class MyApp extends StatelessWidget {
   Widget build(BuildContext context) {
     return MaterialApp(
       title: 'Flutter Demo',
-      theme: TekThemes.light,
-      home: const MyHomePage(title: 'Flutter Demo Home Page'),
+      theme: TekThemes.light.copyWith(
+        textTheme: ThemeData().textTheme,
+      ),
+      home: const DataTableBasicUsageExampleWidget(),
     );
   }
 }
@@ -40,13 +42,52 @@ class DemoDataTableModel {
       };
 }
 
+const String _userNameKey = 'USERNAME_KEY';
 const String _genderKey = 'GENDER_KEY';
-const String _userNameKey = 'USER_NAME_KEY';
 const String _emailKey = 'EMAIL_KEY';
 const String _actionKey = 'ACTION_KEY';
 
-List<DataTableReorderableColumn<DemoDataTableModel>> _columnsDemoDataTable = [
-  DataTableReorderableColumn(
+List<DataTableColumn<DemoDataTableModel>> _columnsDemoDataTable = [
+  DataTableColumn(
+    key: _userNameKey,
+    name: 'UserName',
+    minWidth: 200,
+    flex: 2,
+    customizeItemWidget: (
+      context,
+      value,
+      rowData,
+      columnKey,
+      columnName,
+      width,
+      showOnScreens,
+    ) {
+      return Align(
+        alignment: Alignment.center,
+        child: Row(
+          children: [
+            Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                Text(
+                  rowData.userName,
+                  style: const TextStyle(
+                    fontWeight: FontWeight.bold,
+                  ),
+                ),
+                TekVSpace.mainSpace,
+                TekTags.info(
+                  'ID: ${rowData.id}',
+                  isOutlined: true,
+                ),
+              ],
+            )
+          ],
+        ),
+      );
+    },
+  ),
+  DataTableColumn(
     key: _genderKey,
     name: 'Gender',
     minWidth: 100,
@@ -73,30 +114,7 @@ List<DataTableReorderableColumn<DemoDataTableModel>> _columnsDemoDataTable = [
       );
     },
   ),
-  DataTableReorderableColumn(
-    key: _userNameKey,
-    name: 'User name',
-    minWidth: 200,
-    flex: 2,
-    customizeItemWidget: (
-        context,
-        value,
-        rowData,
-        columnKey,
-        columnName,
-        width,
-        showOnScreens,
-        ) {
-      return Align(
-        alignment: Alignment.center,
-        child: TekTypography(
-          text: rowData.userName,
-          fontWeight: FontWeight.w600,
-        ),
-      );
-    },
-  ),
-  DataTableReorderableColumn(
+  DataTableColumn(
     key: _emailKey,
     name: 'Email',
     minWidth: 200,
@@ -118,7 +136,7 @@ List<DataTableReorderableColumn<DemoDataTableModel>> _columnsDemoDataTable = [
       );
     },
   ),
-  DataTableReorderableColumn(
+  DataTableColumn(
     key: _actionKey,
     name: 'Action',
     minWidth: 120,
@@ -155,84 +173,66 @@ List<DataTableReorderableColumn<DemoDataTableModel>> _columnsDemoDataTable = [
 const List<DemoDataTableModel> _dataSourcesDemoDataTable = [
   DemoDataTableModel(
     id: 1,
-    userName: 'Name of user 1',
+    userName: 'Jerry Ward',
     gender: 'male',
     email: 'jerry.ward@example.com',
   ),
   DemoDataTableModel(
     id: 2,
-    userName: 'Name of user 2',
+    userName: 'Cariana Sales',
     gender: 'female',
     email: 'cariana.sales@example.com',
   ),
   DemoDataTableModel(
     id: 3,
-    userName: 'Name of user 3',
+    userName: 'Kenzo Gautier',
     gender: 'male',
     email: 'kenzo.gautier@example.com',
   ),
   DemoDataTableModel(
     id: 4,
-    userName: 'Name of user 4',
+    userName: 'Arnoldo Toro',
     gender: 'male',
     email: 'arnoldo.toro@example.com',
   ),
   DemoDataTableModel(
     id: 5,
-    userName: 'Name of user 5',
+    userName: 'Claudia Velasco',
     gender: 'female',
     email: 'claudia.velasco@example.com',
   ),
   DemoDataTableModel(
     id: 6,
-    userName: 'Name of user 6',
+    userName: 'Taahira Salian',
     gender: 'male',
     email: 'taahira.salian@example.com',
   ),
   DemoDataTableModel(
     id: 7,
-    userName: 'Name of user 7',
+    userName: 'Sophie Lévesque',
     gender: 'female',
     email: 'sophie.levesque@example.com',
   ),
   DemoDataTableModel(
     id: 8,
-    userName: 'Name of user 8',
-    gender: 'female',
-    email: 'angela.gomez@example.com',
-  ),
-  DemoDataTableModel(
-    id: 9,
-    userName: 'Name of user 9',
-    gender: 'male',
-    email: 'taahira.salian@example.com',
-  ),
-  DemoDataTableModel(
-    id: 10,
-    userName: 'Name of user 10',
-    gender: 'female',
-    email: 'sophie.levesque@example.com',
-  ),
-  DemoDataTableModel(
-    id: 11,
-    userName: 'Name of user 11',
+    userName: 'Angela Gómez',
     gender: 'female',
     email: 'angela.gomez@example.com',
   ),
 ];
 
-class MyHomePage extends StatefulWidget {
-  const MyHomePage({super.key, required this.title});
-
-  final String title;
+class DataTableBasicUsageExampleWidget extends StatefulWidget {
+  const DataTableBasicUsageExampleWidget({super.key});
 
   @override
-  State<MyHomePage> createState() => _MyHomePageState();
+  State<DataTableBasicUsageExampleWidget> createState() =>
+      _DataTableBasicUsageExampleWidgetState();
 }
 
-class _MyHomePageState extends State<MyHomePage> {
-  final TekDataTableReorderableController<DemoDataTableModel> _controller =
-      TekDataTableReorderableController<DemoDataTableModel>();
+class _DataTableBasicUsageExampleWidgetState
+    extends State<DataTableBasicUsageExampleWidget> {
+  final TekDataTableController<DemoDataTableModel> _controller =
+      TekDataTableController<DemoDataTableModel>();
 
   @override
   void initState() {
@@ -253,28 +253,32 @@ class _MyHomePageState extends State<MyHomePage> {
   }
 
   @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text(
-          'Test Reorderable Table',
-          style: context.theme.appBarTheme.titleTextStyle,
-        ),
-      ),
-      body: SingleChildScrollView(
-        padding: EdgeInsets.all(TekSpacings().mainSpacing),
-        child: TekDataTableReorderable<DemoDataTableModel>(
+  Widget build(BuildContext context) => Scaffold(
+        body: TekDataTable<DemoDataTableModel>(
           tableColumns: _columnsDemoDataTable,
           controller: _controller,
           handleChangeData: _handleChangeData,
-          rowOption: TekDataTableReorderableRowOption(
+          additionColumns: const [
+            TekDataTableAdditionColumn.showMore,
+          ],
+          optionUI: const TekDataTableOptionUI(
+            fixTableInAScreen: true,
+          ),
+          rowOption: TekDataTableRowOption(
+            bordered: false,
             paddingOfRowItem: EdgeInsets.all(TekSpacings().mainSpacing),
           ),
-          onReorder: (oldIndex, newIndex){
-            _controller.reorderRow(oldIndex, newIndex);
-          },
+          showerMoreContentIntoRowWidget: (rowData) => Padding(
+            padding: const EdgeInsets.all(8.0),
+            child: Row(
+              children: [
+                TekTags.info(
+                  'ID: ${rowData.id}',
+                  isOutlined: true,
+                ),
+              ],
+            ),
+          ),
         ),
-      ),
-    );
-  }
+      );
 }