+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.6.12
 * Edit `hintText` for `TekInputDropdown`.
+* Adding `TekDataTableAdditionColumn.showMore` to has icon expanded or collapsed in `TekDataTable`.
 ## 1.6.10
 * Add `colorFilter` to `TekSvgImage`.
 ## 1.6.9